fix: Add newline at end of file
Discovered an oddity with how Atom's automatic adding of the newline at
the end of the file means git/GitHub does not pick up the newline as
existing. As a result, while the file passes a pycodestyle check on my
local computer, it fails if the file saved to git/GitHub is checked as
the newline is gone. Have disabled Atom's helper package and trying to
add newline manually to file.